requirements.txt
Django==1.10.6
django-bootstrap3==8.2.1
django-leaflet==0.21.0
olefile==0.44
Pillow==4.0.0
psycopg2==2.7
